requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
async-timeout==4.0.3
asyncpg==0.28.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
dill==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==2.0.2
h11==0.14.0
humanize==4.8.0
idna==3.4
isort==5.12.0
kombu==5.3.2
lazy-object-proxy==1.9.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
passlib==1.7.4
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
pylint==2.17.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomlkit==0.12.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
